round 12: de-unsafe pass — 250 fns reverted from unsafe fn to safe pub fn(&T/NonNull)
NEVER add unsafe fn for not_unsafe_ptr_arg_deref. Strategy:
- Pure-Rust internal helpers: *mut T → &mut T / NonNull<T>; delete # Safety,
  body deref → field access, update callers.
- FFI callback shims (C ABI fn-ptr slots): keep pub fn(*mut T), deref once at
  top with SAFETY comment (the lint is over-strict here — the *mut sig is
  forced by the C ABI, the internal deref is the only unsafe op).
- extern "C" called from C++: keep *mut sig, top-of-body unsafe { &mut *p }
  with SAFETY: C++ never passes null.

Net: -424 unsafe blocks vs main (12,456 → 12,032).

Also: multi-target collector (collect-all-targets.sh + edit-round.workflow.ts),
miri.yml workflow, clippy::all stripped from 67 crates, codegen template
fixes (generate-js2native.ts unused import allow).

+++ src/jsc/RuntimeTranspilerCache.rs
@@ -385,7 +385,7 @@ impl Entry {
                 for v in vecs {
                     debug_assert!(v.len > 0);
                     // `uv_buf_t::len` is `ULONG` (u32) on Windows, `usize` on POSIX.
-                    total += v.len;
+                    total += v.len as usize;
                 }
                 debug_assert!(end_position == total);
             }
